+++ NEWS.md
@@ -1,4 +1,4 @@
-# sjmisc 1.8.0-5
+# sjmisc 2.0.0
 
 ## General
 
@@ -6,6 +6,7 @@
 * All statistical functions have been removed and are now in a separate package, [sjstats](https://cran.r-project.org/package=sjstats).
 * Removed some S3-methods for `labelled`-class, as these are now provided by the haven-package.
 * Functions no longer check input for type `matrix`, to avoid conflicts with scaled vectors (that were recognized as matrix and hence treated as data frame).
+* `table*, exclude = NULL)` was changed to `table*, useNA = "always")`, because of planned changes in upcoming R version 3.4.
 
 ## New functions
 
@@ -18,6 +19,7 @@
 
 * `str_contains` gets a `switch` argument to switch the role of `x` and `pattern`.
 * `word_wrap` coerces vectors to character if necessary.
+* `to_label` gets a `var.label` and `drop.levels` argument, and now preserves variable labels by default.
 
 ## Bug fixes